Add 32-bit unconditional jump instruction
This instruction is part of the cpu=v4 set of instructions
recently added, and specified in
https://www.ietf.org/archive/id/draft-ietf-bpf-isa-00.html#jump-instructions

Fixes #532

+++ src/asm_syntax.hpp
@@ -336,7 +336,8 @@ using InstructionSeq = std::vector<LabeledInstruction>;
 #define DECLARE_EQ1(T, f1) \
     inline bool operator==(T const& a, T const& b) { return a.f1 == b.f1; }
 
-using pc_t = uint16_t;
+// cpu=v4 supports 32-bit PC offsets so we need a large enough type.
+using pc_t = size_t;
 
 // Helpers:
